Added a method to check uniqueness without additional data structure
Using an integer value to represent 32 bits and assuming a simple a-z
input only, we can set the bit of the corresponding character after
subtracting the ASCII value of 'a' and storing it in the integer value.
If we encounter a repeated character, the get will return a positive
value.